add a cast as a hint to the compiler which std::log() to use - the single or the double precision version. I chose to go with the double precision.

+++ dev/apps/swrx/gpsSim.cpp
@@ -58,7 +58,7 @@ using namespace std;
 
 /* exp10() is a gnu specific extensions */
 #ifndef exp10
-#define exp10(x) (exp((x)*log(10)))
+#define exp10(x) (exp((x)*log((long double)10)))
 #endif
 
 class GpsSim : public BasicFramework
